update code snippet order
